requirements.txt
streamlit
openai
python-dotenv
chromadb
PyPDF2